requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
aiofiles
python-multipart
